phantom effect code duplication
