+++ Mage.Sets/src/mage/cards/p/PhantomFlock.java
@@ -27,34 +27,27 @@
  */
 package mage.cards.p;
 
-import java.util.UUID;
 import mage.MageInt;
-import mage.abilities.Ability;
 import mage.abilities.common.EntersBattlefieldAbility;
 import mage.abilities.common.SimpleStaticAbility;
-import mage.abilities.effects.PreventionEffectImpl;
+import mage.abilities.effects.PhantomPreventionEffect;
 import mage.abilities.effects.common.counter.AddCountersSourceEffect;
 import mage.abilities.keyword.FlyingAbility;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.CardType;
-import mage.constants.Duration;
-import mage.constants.PhaseStep;
 import mage.constants.Zone;
 import mage.counters.CounterType;
-import mage.game.Game;
-import mage.game.events.GameEvent;
-import mage.game.permanent.Permanent;
-import mage.game.turn.Step;
+
+import java.util.UUID;
 
 /**
- *
  * @author emerald000
  */
 public class PhantomFlock extends CardImpl {
 
     public PhantomFlock(UUID ownerId, CardSetInfo setInfo) {
-        super(ownerId,setInfo,new CardType[]{CardType.CREATURE},"{3}{W}{W}");
+        super(ownerId, setInfo, new CardType[]{CardType.CREATURE}, "{3}{W}{W}");
         this.subtype.add("Bird");
         this.subtype.add("Soldier");
         this.subtype.add("Spirit");
@@ -63,12 +56,12 @@ public class PhantomFlock extends CardImpl {
 
         // Flying
         this.addAbility(FlyingAbility.getInstance());
-        
+
         // Phantom Flock enters the battlefield with three +1/+1 counters on it.
         this.addAbility(new EntersBattlefieldAbility(new AddCountersSourceEffect(CounterType.P1P1.createInstance(3)), "with three +1/+1 counters on it"));
-        
+
         // If damage would be dealt to Phantom Flock, prevent that damage. Remove a +1/+1 counter from Phantom Flock.
-        this.addAbility(new SimpleStaticAbility(Zone.BATTLEFIELD, new PhantomFlockPreventionEffect()));
+        this.addAbility(new SimpleStaticAbility(Zone.BATTLEFIELD, new PhantomPreventionEffect()));
     }
 
     public PhantomFlock(final PhantomFlock card) {
@@ -79,72 +72,4 @@ public class PhantomFlock extends CardImpl {
     public PhantomFlock copy() {
         return new PhantomFlock(this);
     }
-}
-
-class PhantomFlockPreventionEffect extends PreventionEffectImpl {
-
-    // remember turn and phase step to check if counter in this step was already removed
-    private int turn = 0;
-    private Step combatPhaseStep = null;
-
-    PhantomFlockPreventionEffect() {
-        super(Duration.WhileOnBattlefield);
-        staticText = "If damage would be dealt to {this}, prevent that damage. Remove a +1/+1 counter from {this}";
-    }
-
-    PhantomFlockPreventionEffect(final PhantomFlockPreventionEffect effect) {
-        super(effect);
-        this.turn = effect.turn;
-        this.combatPhaseStep = effect.combatPhaseStep;
-    }
-
-    @Override
-    public PhantomFlockPreventionEffect copy() {
-        return new PhantomFlockPreventionEffect(this);
-    }
-
-    @Override
-    public boolean apply(Game game, Ability source) {
-        return true;
-    }
-
-    @Override
-    public boolean replaceEvent(GameEvent event, Ability source, Game game) {
-        preventDamageAction(event, source, game);
-
-        Permanent permanent = game.getPermanent(source.getSourceId());
-        if (permanent != null) {
-            boolean removeCounter = true;
-            // check if in the same combat damage step already a counter was removed
-            if (game.getTurn().getPhase().getStep().getType().equals(PhaseStep.COMBAT_DAMAGE)) {
-                if (game.getTurnNum() == turn
-                        && game.getTurn().getStep().equals(combatPhaseStep)) {
-                    removeCounter = false;
-                } else {
-                    turn = game.getTurnNum();
-                    combatPhaseStep = game.getTurn().getStep();
-                }
-            }
-
-            if(removeCounter && permanent.getCounters(game).containsKey(CounterType.P1P1)) {
-                StringBuilder sb = new StringBuilder(permanent.getName()).append(": ");
-                permanent.removeCounters(CounterType.P1P1.createInstance(), game);
-                sb.append("Removed a +1/+1 counter ");
-                game.informPlayers(sb.toString());
-            }
-        }
-
-        return false;
-    }
-
-    @Override
-    public boolean applies(GameEvent event, Ability source, Game game) {
-        if (super.applies(event, source, game)) {
-            if (event.getTargetId().equals(source.getSourceId())) {
-                return true;
-            }
-        }
-        return false;
-    }
-
 }
